Move wds-status
